+++ docker/test-compose.yaml
@@ -14,5 +14,7 @@ services:
     command: bash -c 'yarn test-e2e'
     environment:
       E2E_FRONTEND_URL: ${TEST_URL:-https://localhost:8080}
+      FEATURE_TOGGLE_API_URL: ${FEATURE_TOGGLE_API_URL:-http://localhost:4002}
+      IDAM_API_URL: ${IDAM_API_URL:-http://localhost:4501}
     volumes:
       - "../.:/tests"
