+++ package.json
@@ -26,9 +26,9 @@
     "test-unit": "NODE_ENV=testing LOG_LEVEL=OFF NODE_PATH=. mocha './app/**/*.test.js' --reporter spec --recursive -t 3000",
     "test-unit-ide": "NODE_ENV=testing LOG_LEVEL=OFF NODE_PATH=. mocha $NODE_DEBUG_OPTION './app/**/*.test.js' --reporter spec --recursive",
     "test-validation": "NODE_ENV=testing LOG_LEVEL=OFF NODE_PATH=. mocha test/validation/ --reporter spec --recursive --timeout 15000",
-    "test-e2e": "NODE_ENV=testing LOG_LEVEL=OFF NODE_PATH=. codeceptjs run -c test/end-to-end/ --steps --grep @overnight --invert",
+    "test-e2e": "NODE_ENV=testing LOG_LEVEL=OFF NODE_PATH=. codeceptjs run -c codecept.conf.js --steps --grep @overnight --invert",
     "pretest-e2e": "sleep 10",
-    "test-overnight-e2e": "NODE_ENV=testing LOG_LEVEL=OFF NODE_PATH=. codeceptjs run -c test/end-to-end/ --steps",
+    "test-overnight-e2e": "NODE_ENV=testing LOG_LEVEL=OFF NODE_PATH=. codeceptjs run codecept.conf.js --steps",
     "test-crossbrowser-e2e": "NODE_ENV=testing LOG_LEVEL=OFF NODE_PATH=. codeceptjs run -c test/end-to-end/saucelabs.conf.js --steps",
     "test-all": "NODE_ENV=testing LOG_LEVEL=OFF yarn run test-coverage && yarn run test-a11y && yarn run test-nsp",
     "sonar-scanner": "node_modules/sonar-scanner/bin/sonar-scanner",
